bench: Don't drop the output data
Since in the current benchmarks the output data is discarded, compiler
is smart enough to eliminate all data copying (both in to the buffer,
and out of the buffer), and roll over the original buffer only.

With this change the overhead of coping things is being measured as
well. I still wonder if both copyings are being done, or just the one
out of the buffer.

In the real code, at least the out of the buffer copying would be
unavoidable, as unless something can be done immediately with that data,
it has to be copied somewhere, eg. to send for processing to another
thread.